+++ tests/unit/precice.py
@@ -62,3 +62,9 @@ class Participant:
             return self.read_write_scalar_buffer
         elif data_name == "macro-vector-data":
             return self.read_write_vector_buffer
+
+    def start_profiling_section(self, section_name):
+        pass
+
+    def stop_last_profiling_section(self):
+        pass